requirements.txt
fastapi[all]==0.115.12
uvicorn[standard]==0.34.0
aiohttp[speedups]==3.11.16
SQLAlchemy==2.0.40
celery==5.5.0
alembic==1.15.2
pydantic==2.11.2
pydantic-settings==2.8.1
ipython==9.0.2
jedi==0.19.2
ipdb==0.13.13
PyJWT==2.10.1
psycopg2-binary==2.9.10
asyncpg==0.30.0
orjson==3.10.16
gunicorn==23.0.0
python-jose[cryptography]==3.4.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
email_validator==2.2.0
python-multipart==0.0.20
aiosmtplib==4.0.0
Jinja2==3.1.6
redis==5.2.1
requests==2.32.3
openai==1.70.0
